Remove testutils package
The RunningOnCircleCI, IsRunningInContainer test-utils were no longer
needed because CI is not running on CircleCI in this repo, so we don't
need these utilities.

Given that the `_test.go` file is only built on Linux, I moved the
remaining SetupTestOSContext utility to the _test.go file itself as
a helper, and removed the testutils package.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>